area function and test

+++ structs_methods_interfaces/perimeter.go
@@ -1,6 +1,6 @@
 package main
 
-func Perimeter(width, height int) int {
-    perimeter := width * 2 + height * 2
-    return perimeter
+func Perimeter(width, height float64) float64 {
+	perimeter := width*2 + height*2
+	return perimeter
 }
